% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{passo}
\alias{passo}
\title{Format a short bullet-point calculation step}
\usage{
passo(...)
}
\arguments{
\item{...}{Passed to \code{\link[=sprintf]{sprintf()}} to build the message.}
}
\value{
A character scalar.
}
\description{
Format a short bullet-point calculation step
}
\keyword{internal}
