% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cap07-tamanho-guarda-chuva.R
\name{razao_tamanho_R}
\alias{razao_tamanho_R}
\title{Sample size to estimate the population ratio R}
\usage{
razao_tamanho_R(
  sr2,
  N,
  X_barra,
  d,
  conf = 0.95,
  arredondar = c("cima", "proximo"),
  ...
)
}
\arguments{
\item{sr2}{Variance of the ratio estimator (population value, or a
pilot-sample estimate from \code{\link[=razao_variancia_piloto]{razao_variancia_piloto()}}).}

\item{N}{Population size.}

\item{X_barra}{Population mean of the auxiliary variable (\code{X/N}).}

\item{d}{Tolerable margin of error.}

\item{conf}{Confidence level, default \code{0.95}.}

\item{arredondar}{Either \code{"cima"} (default) or \code{"proximo"} -- see
\code{\link[=aas_tamanho_media]{aas_tamanho_media()}}.}

\item{...}{Ignored (see package overview on tolerated extra arguments).}
}
\value{
A \code{sampling_result} object with the required sample size.
}
\description{
Computes the sample size needed to estimate the population ratio \code{R}
with a given margin of error (Eq. 7.41 of the course notes).
}
\examples{
razao_tamanho_R(sr2 = 7.7853, N = 1500, X_barra = 10, d = 0.05)
}
