% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validar_proporcao}
\alias{validar_proporcao}
\title{Validate a proportion (must be between 0 and 1, inclusive)}
\usage{
validar_proporcao(p, nome = "p")
}
\arguments{
\item{p}{Value to validate.}

\item{nome}{Character. Name of the argument, used in error messages.}
}
\value{
\code{p}, invisibly, if valid; otherwise raises an error.
}
\description{
Validate a proportion (must be between 0 and 1, inclusive)
}
\keyword{internal}
